refactor: move token out of trait, add custom constrains
